+++ business/manager/master/src/main/java/sharemer/business/manager/master/service/impl/MusicListServiceImpl.java
@@ -45,7 +45,7 @@ public class MusicListServiceImpl implements MusicListService {
     }
 
     @Override
-    public MusicList getOneByWyId(Integer wy_id) {
+    public MusicList getOneByWyId(Long wy_id) {
         return this.musicListMapper.getOneByWyId(wy_id);
     }
